TEST': fixed game logical

+++ src/main/scala/controller/states/commonPlayer/WeaponState.scala
@@ -6,12 +6,9 @@ import controller.GameController
 import model.general.schedule.TurnSchedule
 import model.spell.Spell
 import model.weapons.Weapon
-class WeaponState (private val src: GameUnit, private var spell: Option[Spell],val people : TurnSchedule) extends AGameState {
+class WeaponState (private val src: GameUnit,val people : TurnSchedule) extends AGameState {
   var choice : Int = 0
-  def this(src:GameUnit,t:TurnSchedule) = {
-    this(src,None,t)
-  }
-
+  src.dropWeapon()
   private var selected: Option[Weapon] = None
 
   override def notify(controller: GameController) = {
@@ -28,10 +25,12 @@ class WeaponState (private val src: GameUnit, private var spell: Option[Spell],v
       }
       else{
         val w = src.weapons()(choice-1)
-        src.putWeapon(w)
-        selected = Some(w)
-        people.deletePlayer(src)
-        people.addPlayer(src)
+        if (w.getOwner().isEmpty){
+          src.putWeapon(w)
+          selected = Some(w)
+          people.deletePlayer(src)
+          people.addPlayer(src)
+        } else choice = 6
       }
     }
     catch {
@@ -40,13 +39,19 @@ class WeaponState (private val src: GameUnit, private var spell: Option[Spell],v
   }
 
   override def update(controller: GameController): Unit = {
-    choice match {
-      case 0 => controller.state = new ActionState (src, people)
-      case 1 => controller.state = new TargetState (src, people)
-      case 2 => controller.state = new TargetState (src, people)
-      case 3 => controller.state = new TargetState (src, people)
-      case 4 => controller.state = new WeaponState (src, people)
-      case 5 => controller.state = new WeaponState (src, people)
+    if (src.arma.isEmpty){
+      controller.state = new WeaponState(src,people)
+    }
+    else {
+      choice match {
+        case 0 => controller.state = new ActionState (src, people)
+        case 1 => controller.state = new TargetState (src, people)
+        case 2 => controller.state = new TargetState (src, people)
+        case 3 => controller.state = new TargetState (src, people)
+        case 4 => controller.state = new WeaponState (src, people)
+        case 5 => controller.state = new WeaponState (src, people)
+        case 6 => controller.state = new WeaponState(src,people)
+      }
     }
   }
